% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_absolute_QALYs.R
\name{calculate_absolute_QALYs}
\alias{calculate_absolute_QALYs}
\title{Calculate Absolute Quality-Adjusted Life Years (QALYs).}
\usage{
calculate_absolute_QALYs(
  baseline_health_ = input_data_mQALE$`Baseline health`,
  mortality_elasticity_ = input_data_mQALE$`Mortality elasticity`,
  option_ = "estimated_mortality_elasticity",
  imd_population_ = CCG_IMD_population_2019,
  provider_ = "CCG"
)
}
\arguments{
\item{baseline_health_}{Numeric vector specifying the baseline health per
quintile. Default value from PAPER 1 representing EQ5D-5L results.}

\item{mortality_elasticity_}{Numeric vector specifying the mortality
elasticity per deprivation quintile.}

\item{option_}{Character scalar specifying whether to estimate the marginal
QALE based on actual or equal marginal elasticity,
\code{estimated_mortality_elasticity} or \code{equal_mortality_elasticity},
respectively.}

\item{imd_population_}{Dataframe with population distribution by deprivation.}

\item{provider_}{String specifying whether the provider was CCG or ICS?ICB.}
}
\value{
A dataframe.
}
\description{
Estimates the the absolute Quality-Adjusted Life Year (QALY) per
deprivation quintile.
}
\details{
This function takes ... inputs and then ... to estimate the absolute
QALYs.
}
