% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_average_deaths.R
\name{calculate_average_deaths}
\alias{calculate_average_deaths}
\title{Calculate Average Number of Deaths Prevented by Deprivation Quintile.}
\usage{
calculate_average_deaths(
  total_lives_saved_ = UnmetNeeds::calculate_total_deaths()$data,
  imd_population_ = CCG_IMD_population_2019
)
}
\arguments{
\item{total_lives_saved_}{Dataframe or data table representing the outputs of
the \link[UnmetNeeds]{calculate_absolute_QALYs}.}

\item{imd_population_}{Dataframe with population distribution by deprivation.}
}
\value{
A dataframe.
}
\description{
Estimates the the average Quality-Adjusted Life Year (QALY) per
deprivation quintile.
}
\details{
This function takes ... inputs and then ... to estimate the average
QALYs.
}
\examples{
\dontrun{
average_deaths <- calculate_average_deaths(
  total_lives_saved_ = calculate_total_deaths(
    mortality_rates_ = input_data_mQALE$`Mortality rate`[1:5],
    mortality_elasticity_ = input_data_mQALE$`Mortality elasticity`,
    equal_mortality_elasticity_ = input_data_mQALE$`Equal mortality elasticity`,
    option_ = "estimated_mortality_elasticity",
    imd_population_ = CCG_IMD_population_2019,
    provider_ = "CCG"
  )$data,
  imd_population_ = CCG_IMD_population_2019
)
average_deaths_equ_elas <- calculate_average_deaths(
  total_lives_saved_ = calculate_total_deaths(
    mortality_rates_ = input_data_mQALE$`Mortality rate`[1:5],
    mortality_elasticity_ = input_data_mQALE$`Mortality elasticity`,
    equal_mortality_elasticity_ = input_data_mQALE$`Equal mortality elasticity`,
    option_ = "equal_mortality_elasticity",
    imd_population_ = CCG_IMD_population_2019,
    provider_ = "CCG"
  )$data,
  imd_population_ = CCG_IMD_population_2019
)
}
}
