% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_average_QALYs.R
\name{calculate_average_QALYs}
\alias{calculate_average_QALYs}
\title{Calculate Average Quality-Adjusted Life Years (QALYs) by Deprivation
Quintile.}
\usage{
calculate_average_QALYs(
  absolute_QALYs_ = UnmetNeeds::calculate_absolute_QALYs()$data,
  imd_population_ = CCG_IMD_population_2019
)
}
\arguments{
\item{absolute_QALYs_}{Dataframe or data table representing the outputs of
the \link[UnmetNeeds]{calculate_absolute_QALYs}.}

\item{imd_population_}{Dataframe with population distribution by deprivation.}
}
\value{
A dataframe.
}
\description{
Estimates the the average Quality-Adjusted Life Year (QALY) per
deprivation quintile.
}
\details{
This function takes ... inputs and then ... to estimate the average
QALYs.
}
