% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_marginal_QALYs.R
\name{calculate_marginal_QALYs}
\alias{calculate_marginal_QALYs}
\title{Calculate the marginal QALYs}
\usage{
calculate_marginal_QALYs(
  target_maximum_health_ = input_data_mQALE$`Target maximum QALE`,
  baseline_health_ = input_data_mQALE$`Baseline health`,
  mortality_elasticity_ = input_data_mQALE$`Mortality elasticity`,
  option_ = "estimated_mortality_elasticity",
  expenditure_change_ = 1
)
}
\arguments{
\item{target_maximum_health_}{Numeric vector specifying the target maximum QALE
for low (GBD study), mid and high (maximum lifespan).}

\item{baseline_health_}{Numeric vector specifying the baseline health per
quintile. Default value from PAPER 1 representing EQ5D-5L results.}

\item{mortality_elasticity_}{Numeric vector specifying the mortality
elasticity per deprivation quintile.}

\item{option_}{Character scalar specifying whether to estimate the marginal
QALE based on actual or equal marginal elasticity,
\code{estimated_mortality_elasticity} or \code{equal_mortality_elasticity},
respectively.}

\item{expenditure_change_}{Numeric specifying the percentage change for which
the QALY gain/loss is to be estimated.}
}
\value{
A dataframe.
}
\description{
Estimates the per deprivation quintile marginal Quality-Adjusted
Life Years (QALYs) from the Quality-Adjusted Life Expectancy (QALE).
}
\details{
This function takes ... inputs and then ... to estimate the marginal
QALE per deprivation quintile and then the marginal QALYs.
}
