% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_total_deaths.R
\name{calculate_total_deaths}
\alias{calculate_total_deaths}
\title{Calculate the Total Number of Deaths prevented by healthcare provider}
\usage{
calculate_total_deaths(
  mortality_rates_,
  mortality_elasticity_,
  equal_mortality_elasticity_,
  option_ = "estimated_mortality_elasticity",
  imd_population_,
  provider_ = "CCG"
)
}
\arguments{
\item{mortality_rates_}{Numeric vector specifying the mortality rates per
100,000 population per deprivation quintile.}

\item{mortality_elasticity_}{Numeric vector specifying the mortality
elasticity per deprivation quintile.}

\item{equal_mortality_elasticity_}{Numeric scalar defining the value to be
used when the argument \code{option_} is set to \code{equal_mortality_elasticity}.}

\item{option_}{Character scalar specifying whether to estimate the marginal
QALE based on actual or equal marginal elasticity,
\code{estimated_mortality_elasticity} or \code{equal_mortality_elasticity},
respectively.}

\item{imd_population_}{Dataframe with population distribution by deprivation.}

\item{provider_}{String specifying whether the provider was CCG or ICS?ICB.}
}
\value{
A list containing dataframe(s).
}
\description{
Estimates the total number of deaths per deprivation quintile in
each geographical area with a designated healthcare provider.
}
\details{
This function takes ... inputs and then ... to estimate the total
number of deaths.
}
\examples{
\dontrun{
total_lives_saved <- calculate_total_deaths(
  mortality_rates_ = input_data_mQALE$`Mortality rate`[1:5],
  mortality_elasticity_ = input_data_mQALE$`Mortality elasticity`,
  equal_mortality_elasticity_ = input_data_mQALE$`Equal mortality elasticity`,
  option_ = "estimated_mortality_elasticity",
  imd_population_ = CCG_IMD_population_2019,
  provider_ = "CCG"
)
total_lives_saved_equ_elas <- calculate_total_deaths(
  mortality_rates_ = input_data_mQALE$`Mortality rate`[1:5],
  mortality_elasticity_ = input_data_mQALE$`Mortality elasticity`,
  equal_mortality_elasticity_ = input_data_mQALE$`Equal mortality elasticity`,
  option_ = "equal_mortality_elasticity",
  imd_population_ = CCG_IMD_population_2019,
  provider_ = "CCG"
)
}
}
