% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{create_map_plot}
\alias{create_map_plot}
\title{Create a landmap plot}
\usage{
create_map_plot(
  sp_df_,
  value_df_,
  var_nm_,
  value_year_ = NA,
  value_digits_ = 0
)
}
\arguments{
\item{sp_df_}{A SpatialPointsDataFrame defining the boundaries within
the land map.}

\item{value_df_}{A dataframe containing average values and scaled average
values (between 0 and 1) for each area within the land map.}

\item{var_nm_}{A string specifying the name of the variable being displayed.}

\item{value_year_}{(Optional) A numeric specifying the year of the variable
being displayed.}

\item{value_digits_}{A numeric scalar specifying the number of digits to
which the outcome in the data table or dataframe \code{value_df_} is to be rounds.}
}
\value{
A Leaflet object.
}
\description{
Creates a landmap plot of average values.
}
\details{
Creates a landmap plot using Leaflet displaying the average values
given by local authority in a dataframe
}
\examples{
\dontrun{
map_data = sp_df_2023 |>
  slot("data") |>
  subset(x = _, select = ICB23CD) |>
  cbind(
  "Value" = runif(nrow(sp_df_2023@data))
  )

create_map_plot(sp_df_ = sp_df_2023,
                value_df_ = map_data,
                var_nm_ = "QALYs",
                value_year_ = 2023)

map_data2 = sp_df_2018 |>
  slot("data") |>
  subset(x = _, select = ccg18cd) |>
  cbind(
  "Value" = runif(nrow(sp_df_2018@data))
  )

create_map_plot(sp_df_ = sp_df_2018,
                value_df_ = map_data2,
                var_nm_ = "QALYs",
                value_year_ = 2018)
}

}
