% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{error_on_missing_name}
\alias{error_on_missing_name}
\title{Check shinyapps.io Account Secrets}
\usage{
error_on_missing_name(name_)
}
\arguments{
\item{name_}{The name of the secret/variable to find}
}
\value{
Unquoted value of the secret
}
\description{
a function to stop the script when one of the variables cannot be found.
and to strip quotation marks from the secrets when you supplied them.
}
\examples{
\dontrun{
error_on_missing_name(name_ = "app")
}
}
