% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_average_deaths.R
\name{plot_average_deaths}
\alias{plot_average_deaths}
\title{Plot the average deaths by deprivation quintile}
\usage{
plot_average_deaths(average_deaths_)
}
\arguments{
\item{average_deaths_}{An object representing the outputs of the function
\link[UnmetNeeds]{calculate_average_deaths}.}
}
\value{
A list containing ggplot2 objects.
}
\description{
Plot the average deaths by deprivation quintile
}
\details{
Plot the average deaths.
}
\examples{
\dontrun{
average_deaths_bar_chart <- plot_average_deaths()
}
}
