% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_average_QALYs.R
\name{plot_average_QALYs}
\alias{plot_average_QALYs}
\title{Plot the average change in Quality-Adjusted Life Years (QALYs)}
\usage{
plot_average_QALYs(average_QALYs_)
}
\arguments{
\item{average_QALYs_}{An object representing the outputs of the function
\link[UnmetNeeds]{calculate_average_QALYs}.}
}
\value{
A list containing ggplot2 objects.
}
\description{
Plot the average change in Quality-Adjusted Life Years (QALYs)
}
\details{
This function takes ... inputs and then ... to plot the average
change in Quality-Adjusted Life Years (QALYs).
}
\examples{
\dontrun{
average_QALYs_change_plot <- plot_average_QALYs()
}
}
