% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_marginal_QALYs.R
\name{plot_marginal_QALYs}
\alias{plot_marginal_QALYs}
\title{Calculate the marginal QALYs}
\usage{
plot_marginal_QALYs(marginal_QALYs_)
}
\arguments{
\item{marginal_QALYs_}{An object representing the outputs of the function
\link[UnmetNeeds]{calculate_marginal_QALYs}.}
}
\value{
A list containing ggplot2 objects.
}
\description{
Estimates the the marginal Quality-Adjusted Life Expectancy
(QALE) per deprivation quintile.
}
\details{
This function takes ... inputs and then ... to estimate the marginal
QALE
}
