% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ShinyNeeds_app.R
\name{run_UnmetNeeds_app}
\alias{run_UnmetNeeds_app}
\title{Run the UnmetNeeds shiny App}
\usage{
run_UnmetNeeds_app(port_ = NULL)
}
\arguments{
\item{port_}{Positive integer, the TCP port that the application should
listen on. See the \code{port} argument in \code{\link[shiny]{runApp}} for
more information.}
}
\value{
Returns nothing, but launches a shiny app in a web browser window.
}
\description{
Run the UnmetNeeds shiny App
}
\examples{
\dontrun{
library(UnmetNeeds)
run_UnmetNeeds_app()
}

}
