% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_absolute_QALYs.R
\name{calculate_absolute_QALYs}
\alias{calculate_absolute_QALYs}
\title{Calculate Absolute Quality-Adjusted Life Years (QALYs)}
\usage{
calculate_absolute_QALYs(
  target_maximum_health_,
  baseline_health_,
  mortality_elasticity_,
  equal_mortality_elasticity_,
  option_ = "estimated_mortality_elasticity",
  imd_population_,
  provider_ = "CCG"
)
}
\arguments{
\item{target_maximum_health_}{Numeric vector specifying the target maximum QALE
for low (GBD study), mid and high (maximum lifespan).}

\item{baseline_health_}{Numeric vector specifying the baseline health per
quintile. Default value from PAPER 1 representing EQ5D-5L results.}

\item{mortality_elasticity_}{Numeric vector specifying the mortality
elasticity per deprivation quintile.}

\item{equal_mortality_elasticity_}{Numeric scalar defining the value to be
used when the argument \code{option_} is set to \code{equal_mortality_elasticity}.}

\item{option_}{Character scalar specifying whether to estimate the marginal
QALE based on actual or equal marginal elasticity,
\code{estimated_mortality_elasticity} or \code{equal_mortality_elasticity},
respectively.}

\item{imd_population_}{Dataframe with population distribution by deprivation.}

\item{provider_}{String specifying whether the provider was CCG or ICS?ICB.}
}
\value{
A list containing dataframe(s).
}
\description{
Estimates the absolute Quality-Adjusted Life Year (QALY) per
deprivation quintile.
}
\details{
This function takes ... inputs and then ... to estimate the absolute
QALYs.
}
\examples{
\dontrun{
absolute_QALYs <- calculate_absolute_QALYs(
  target_maximum_health_ = input_data_mQALE$`Target maximum QALE`,
  baseline_health_ = input_data_mQALE$`Baseline health`[1:5],
  mortality_elasticity_ = input_data_mQALE$`Mortality elasticity`,
  equal_mortality_elasticity_ = input_data_mQALE$`Equal mortality elasticity`,
  option_ = "estimated_mortality_elasticity",
  imd_population_ = CCG_IMD_population_2019,
  provider_ = "CCG"
)
absolute_QALYs_equ_elas <- calculate_absolute_QALYs(
  target_maximum_health_ = input_data_mQALE$`Target maximum QALE`,
  baseline_health_ = input_data_mQALE$`Baseline health`[1:5],
  mortality_elasticity_ = input_data_mQALE$`Mortality elasticity`,
  equal_mortality_elasticity_ = input_data_mQALE$`Equal mortality elasticity`,
  option_ = "equal_mortality_elasticity",
  imd_population_ = CCG_IMD_population_2019,
  provider_ = "CCG"
)
}
}
