% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_average_QALYs.R
\name{calculate_average_QALYs}
\alias{calculate_average_QALYs}
\title{Calculate Average Quality-Adjusted Life Years (QALYs) by Deprivation
Quintile}
\usage{
calculate_average_QALYs(absolute_QALYs_, imd_population_)
}
\arguments{
\item{absolute_QALYs_}{Dataframe or data table representing the outputs of
the \link[UnmetNeeds]{calculate_absolute_QALYs}.}

\item{imd_population_}{Dataframe with population distribution by deprivation.}
}
\value{
A dataframe.
}
\description{
Estimates the the average Quality-Adjusted Life Year (QALY) per
deprivation quintile.
}
\details{
This function takes ... inputs and then ... to estimate the average
QALYs.
}
\examples{
\dontrun{
average_QALYs <- calculate_average_QALYs(
  absolute_QALYs_ = calculate_absolute_QALYs(
    target_maximum_health_ = input_data_mQALE$`Target maximum QALE`,
    baseline_health_ = input_data_mQALE$`Baseline health`[1:5],
    mortality_elasticity_ = input_data_mQALE$`Mortality elasticity`,
    equal_mortality_elasticity_ = input_data_mQALE$`Equal mortality elasticity`,
    option_ = "estimated_mortality_elasticity",
    imd_population_ = CCG_IMD_population_2019,
    provider_ = "CCG"
  )$data[[1]],
  imd_population_ = CCG_IMD_population_2019
)
average_QALYs_equ_elas <- calculate_average_QALYs(
  absolute_QALYs_ = calculate_absolute_QALYs(
    target_maximum_health_ = input_data_mQALE$`Target maximum QALE`,
    baseline_health_ = input_data_mQALE$`Baseline health`[1:5],
    mortality_elasticity_ = input_data_mQALE$`Mortality elasticity`,
    equal_mortality_elasticity_ = input_data_mQALE$`Equal mortality elasticity`,
    option_ = "equal_mortality_elasticity",
    imd_population_ = CCG_IMD_population_2019,
    provider_ = "CCG"
  )$data[[1]],
  imd_population_ = CCG_IMD_population_2019
)
}
}
